vboot/secdata: rewrite rollback_index and centralize reads/writes
In current kernel verification code, secdata reads and writes
are spread throughout the code.  vboot2's design is to use
vb2_context.secdata_* for storing the state of secdata spaces,
and have the caller (depthcharge) read/save this field when
necessary.

Centralize secdata reads/writes into the functions of
secdata_tpm.c, previously known as rollback_index.c.
Functions which directly read/write to the TPM space are modified
to use vb2_secdata_*_get and vb2_secdata_*_set.

The secure spaces get read/flushed by functions in
vboot_api_kernel.c.  These calls and the underlying functions
from secdata_tpm.c will eventually be relocated to depthcharge.

BUG=b:124141368, chromium:972956
TEST=make clean && make runtests
BRANCH=none

+++ firmware/2lib/include/2api.h
@@ -476,10 +476,11 @@ uint32_t vb2api_secdata_kernel_create(struct vb2_context *ctx);
  * and VB2_ERROR_SECDATA_FWMP_INCOMPLETE is returned.  The caller should
  * re-read the returned number of bytes, and call this function again.
  *
- * @param ctx          Context pointer
+ * @param ctx		Context pointer
+ * @param size		Amount of struct which has been read
  * @return VB2_SUCCESS, or non-zero error code if error.
  */
-vb2_error_t vb2api_secdata_fwmp_check(struct vb2_context *ctx, uint32_t *size);
+vb2_error_t vb2api_secdata_fwmp_check(struct vb2_context *ctx, uint8_t *size);
 
 /**
  * Create fresh data in the firmware management parameters (FWMP) space.
@@ -488,7 +489,7 @@ vb2_error_t vb2api_secdata_fwmp_check(struct vb2_context *ctx, uint32_t *size);
  * vb2api_secdata_fwmp_check() (or any other API in this library) fails; that
  * could allow the secure data to be rolled back to an insecure state.
  *
- * @param ctx          Context pointer
+ * @param ctx		Context pointer
  * @return size of created FWMP data in bytes
  */
 uint32_t vb2api_secdata_fwmp_create(struct vb2_context *ctx);
